moved as much as possible out of datatypes
